samples: Add a sample for non-flash-source
A sample for non-flash-source on MEC17 EVB. It provides implementation
for Zephyr flash_area_open_custom(), so the right flash map
implementation is used, and MCUboot flash_map_id_get_next() and
flash_map_id_get_current() to prioritize sources. It should show what is
expected from an application to be able to use non-flash sources for
images.

For the I2C device simulation, an Aardvark I2C host adapter is needed,
and a script is provided to make the image loadable via I2C. An
implementation for an "Aardvark driver" on top of I2C is also provided.
Finally, a sample application to be loaded is also available.

For more details on how to build and test the samples, check the
provided README.md.

Signed-off-by: Ederson de Souza <ederson.desouza@intel.com>